requirements.txt
python-dotenv
pycrypto==2.6.1